+++ NEWS.md
@@ -1,4 +1,6 @@
 # 0.2.0 ongoing
+* Added importFrom for `utils`.
+* `CutthroatALF.R`: Modified.  Changed to use `utils::read.table()` rather than `read.table()`.
 
 # 0.1.9 Jul15
 * Removed several URLs that either did not work consistently or have been moved.  Fixed several others.